requirements.txt
aiogram==3.18.0
aiohttp==3.11.13
aiosqlite==0.21.0
greenlet==3.1.1
magic-filter==1.0.12
rich==13.9.4
python-dotenv==1.0.1
SQLAlchemy==2.0.38
